- Resolved issue where flattening a list attempted to cast its contents to integers.

+++ src/main/java/com/patrickangle/commons/util/legacy/ListUtils.java
@@ -214,7 +214,7 @@ public class ListUtils {
             if (item instanceof List<?>) {
                 flattenHelper((List<?>) item, flatList);
             } else {
-                flatList.add((Integer) item);
+                flatList.add(item);
             }
         }
     }
